requirements.txt
python-telegram-bot==22.5
yt-dlp==2025.9.26
requests>=2.31.0
beautifulsoup4>=4.12.2
tqdm>=4.66.1
Flask>=3.0.0
gunicorn>=21.2.0